Let one card weigh more than an int, because a real set does
A line in the log from a real session: "Collation for msh: MSH's collation
could not be read: msh collector, sheet 'foilBoosterfun', card 0368af59:
weight 2274495300 is past what one card can carry". Marvel's Spider-Man
writes its foil odds as exact integer ratios, and one of those numbers is
past what an int holds - so the whole set's collation was refused and every
pack of it opened out of the rarity fallback instead of the real sheets.
Nothing said so except that line.

The sum of a sheet has been a long since it was written, for exactly this
reason - published foil odds come to hundreds of billions - and there was
never a reason for one card on it to be smaller. So a card's weight is a
long now, all the way from the file to the draw.

Counts are unchanged and still bounded: how many cards are in a slot and how
many of a product are in a box are things somebody physically holds, so a
number past an int there is a broken file rather than an unusual set. They
keep the old check, which is now the only caller of it.

The guard is a set file with 2,274,495,300 on one card of a two-card sheet
and 1 on the other, which asserts the light card is still reachable at the
very last point on the sheet - so a reader that cannot hold the heavy number
comes up with the first card every time and fails. Checked by putting the
old cap back and watching it fail.

Gate green: build, 299 game tests, the eight checks.

+++ core/src/main/java/dev/gathering/core/booster/BoosterFallback.java
@@ -48,11 +48,11 @@ public final class BoosterFallback {
             if (rarity == null || printings == null || printings.isEmpty()) {
                 continue;
             }
-            Map<UUID, Integer> weights = new LinkedHashMap<>();
+            Map<UUID, Long> weights = new LinkedHashMap<>();
             for (UUID printing : printings) {
                 // Once each. A card that appears twice in the catalog is still one card in
                 // the set, and counting it twice would make it quietly twice as common.
-                weights.putIfAbsent(printing, 1);
+                weights.putIfAbsent(printing, 1L);
             }
             sheets.put(nameOf(rarity), new BoosterSheet(nameOf(rarity), false, false, false, weights));
         }
