+++ core/src/test/java/dev/gathering/core/booster/CoverageAuditTest.java
@@ -69,7 +69,7 @@ class CoverageAuditTest {
         Map<String, BoosterSheet> sheets = new LinkedHashMap<>();
         sheets.put("common", sheetOf("common", "c1", "c2", "c3"));
         sheets.put("theList", new BoosterSheet(
-                "theList", false, false, false, Map.of(onlyOnTheUnusedSheet, 1)));
+                "theList", false, false, false, Map.of(onlyOnTheUnusedSheet, 1L)));
         BoosterConfig config = new BoosterConfig("tst", "draft", sheets,
                 List.of(new BoosterVariant("plain", 1, Map.of("common", 2))));
 
@@ -222,9 +222,9 @@ class CoverageAuditTest {
     }
 
     private static BoosterSheet sheetOf(String name, String... cards) {
-        Map<UUID, Integer> weights = new LinkedHashMap<>();
+        Map<UUID, Long> weights = new LinkedHashMap<>();
         for (String card : cards) {
-            weights.put(printing(card), 1);
+            weights.put(printing(card), 1L);
         }
         return new BoosterSheet(name, false, false, false, weights);
     }
